1) New test cases added 2) Tests separated for different classes
